refactor: handle validator overflow in monitor header boxes (#954)
